Move IM test boilerplate to beforeEach setup (WIP)
The problem is that one of the tests requires the button to NOT be added
to the setup. The question is how to best achieve this. Some ideas:
- Create multiple setup template variables
  - Having only one test using a given variable is bad
  - Should be a way to create different template variables from a common
  base (TODO: look into whether this is currently possible or if it
  creates problems)
- Do not use the template for that specific test
  - Means SOME boilerplate duplication, but possibly less than any other
  method

It's probably good to have multiple setup template variables anyway,
since half the tests add a callback (or two) to the otherwise similar
setup. This also makes it more helpful to have a way to specify templates
from a common base, ideally without them interfering with each other for
different test cases. I'm not sure whether templates share references to
things, though, which is a problem.